requirements.txt
lyzr-automata
streamlit
requests
time
python-dotenv
smtplib
email